money(double) paso 4-6: nucleo del motor a double (edit atomico)
Migra el tipo de trabajo del dinero de float a double en todo el motor, de una
pasada (no compila hasta migrar el conjunto: el techo lo fija el eslabon mas
estrecho). El consenso ya trabajaba en centimos-long, asi que double solo ensancha
la copia de trabajo/display y sube el techo de exactitud al centimo de ~131072
fichas (float32) a ~9e13.

Interfaces: Player/BotPlayerView/DealerView (getStack/getBet/getBote/pagar/setStack/
setBet/getPagar/setPagar + getBote_total/getApuesta_actual/getUltimo_raise/
getCiega_grande/getCiega_pequena). buyin sigue int; win_per sigue float (%).

Concretos: LocalPlayer/RemotePlayer (campos stack/bet/bote/pagar/call_required/
min_raise + spinner BigDecimal con doubleValue + snapshots all-in). Crupier (8 campos
de dinero + CIEGAS + getters DealerView + calcularBoteParaGanador/splitPotForRunItTwice
sobre PotMath double + settlement con amountToCents(double) + auditor Float[]->Double[]
+ setFloat/getFloat->setDouble/getDouble + recover map (double) + wire INIT/UPDATEBLINDS/
RIT_VOTE_REQ/ACTION + predictPostActionBetCents instanceof Double + action[1] siempre
Double en el bucle). GameFrame (CIEGA_*/BLIND_CAP/ACTIVE_BLIND_STRUCTURE/AUTO_CALL_MAX/
bigBlindForSmallBlind/rebuyHeadroom/setTapeteBote/setTapeteCiegas). BlindStructure
(double[][] + helpers). HandPot (side-pots). CanonicalActionRecord (+amountToCents(double)).
Dialogos de ciegas y display (NewGame/EditBlinds/BlindStructureManager/AutoCallMax/
WaitingRoom/Stats/Balance/TablePanel).

Bot: el dinero se NARROW a float SOLO en el borde de la matematica de sizing/ratios de
la IA (mRatio/spr/EV se quedan float por diseno); la salida getBetSize() vuelve a double
y entra al motor exacta.

Harness qa migrado en lockstep (TestBotPlayer/TestDealer/HeadsUpSimulator/Multiway/
FixedStrategyBot/GameFlowSmoke/BlindStructureTest/RunItTwiceSplitTest).

Garantia: Java no permite estrechamiento implicito double->float, asi que cada recorte
de dinero fue error de compilacion resuelto -> no queda asignacion silenciosa que cape
dinero al techo. El golden de transcripcion da H_final BYTE-IDENTICO (consenso no se
movio por debajo del techo) y GameFlowSmoke conserva fichas sobre miles de manos con el
motor en double. 133/133 verde.

+++ src/main/java/com/tonikelope/coronapoker/bot/context/DealerView.java
@@ -29,15 +29,15 @@ public interface DealerView {
 
     int getStreet();
 
-    float getBote_total();
+    double getBote_total();
 
-    float getApuesta_actual();
+    double getApuesta_actual();
 
-    float getUltimo_raise();
+    double getUltimo_raise();
 
-    float getCiega_grande();
+    double getCiega_grande();
 
-    float getCiega_pequeña();
+    double getCiega_pequeña();
 
     int getConta_bet();
